Completed the function documentation

+++ src/main/java/com/schibsted/spt/data/jstl2/impl/NodeUtils.java
@@ -32,7 +32,7 @@ public class NodeUtils {
       !(value.isObject() && value.size() == 0) &&
       !(value.isTextual() && value.asText().length() == 0) &&
       !(value.isArray() && value.size() == 0) &&
-      !(value.isNumber() && value.intValue() == 0) &&
+      !(value.isNumber() && value.doubleValue() == 0.0) &&
       !value.isNull();
   }
